UserRoles

+++ ArticlesApp/Controllers/ArticlesController.cs
@@ -1,25 +1,41 @@
 ﻿using ArticlesApp.Data;
 using ArticlesApp.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 
 namespace ArticlesApp.Controllers
 {
+    [Authorize]
     public class ArticlesController : Controller
     {
         private readonly ApplicationDbContext db;
-        public ArticlesController(ApplicationDbContext context)
+
+        private readonly UserManager<ApplicationUser> _userManager;
+
+        private readonly RoleManager<IdentityRole> _roleManager;
+
+        public ArticlesController(
+            ApplicationDbContext context,
+            UserManager<ApplicationUser> userManager,
+            RoleManager<IdentityRole> roleManager)
         {
             db = context;
+            _userManager = userManager;
+            _roleManager = roleManager;
         }
 
         // Se afiseaza lista tuturor articolelor din baza de date
         // impreuna cu categoria din care fac parte
         // HttpGet implicit
+        // Pentru fiecare articol se afiseaza si utilizatorul care a postat articolul
+
+        [Authorize(Roles = "User,Editor,Admin")]
         public IActionResult Index()
         {
-            var articles = db.Articles.Include("Category");
+            var articles = db.Articles.Include("Category").Include("User");
 
             // ViewBag.OriceDenumireSugestiva
             ViewBag.Articles = articles;
@@ -34,13 +50,19 @@ namespace ArticlesApp.Controllers
 
         // Se afiseaza un singur articol in functie de id-ul sau 
         // impreuna cu categoria din care face parte
-        // Httpget implicit
+        // HttpGet implicit
+
+        [Authorize(Roles = "User,Editor,Admin")]
         public IActionResult Show(int id) 
         {
             Article article = db.Articles.Include("Category")
+                                         .Include("User")
                                          .Include("Comments")
+                                         .Include("Comments.User")
                                          .Where(art => art.Id == id)
                                          .First();
+
+            SetAccessRights();
 
             // ViewBag.Article = article;
             // ViewBag.Category = article.Category;
@@ -50,11 +72,25 @@ namespace ArticlesApp.Controllers
             return View(article);
         }
 
+        private void SetAccessRights()
+        {
+            ViewBag.AfisareButoane = false;
+
+            if (User.IsInRole("Editor"))
+            {
+                ViewBag.AfisareButoane = true;
+            }
+
+            ViewBag.UserCurent = _userManager.GetUserId(User);
+            ViewBag.EsteAdmin = User.IsInRole("Admin");
+        }
+
         // Adaugarea unui comentariu asociat unui articol in baza de date
         [HttpPost]
         public IActionResult Show([FromForm] Comment comm)
         {
             comm.Date = DateTime.Now;
+            comm.UserId = _userManager.GetUserId(User);
 
             if(ModelState.IsValid)
             {
@@ -65,9 +101,13 @@ namespace ArticlesApp.Controllers
             else
             {
                 Article art = db.Articles.Include("Category")
-                                             .Include("Comments")
-                                             .Where(art => art.Id == comm.ArticleId)
-                                             .First();
+                                         .Include("User")
+                                         .Include("Comments")
+                                         .Include("Comments.User")
+                                         .Where(art => art.Id == comm.ArticleId)
+                                         .First();
+
+                SetAccessRights();
 
                 return View(art);
             }
@@ -77,7 +117,9 @@ namespace ArticlesApp.Controllers
         // Se afiseaza formularul in care se vor completa datele unui articol
         // impreuna cu selectarea categoriei din care face parte articolul
         // HtppGet implicit
+        // Doar utilizatorii cu rolul Editor sau Admin pot adauga articole in platforma
 
+        [Authorize(Roles = "Editor,Admin")]
         public IActionResult New() 
         {
             // var categories = from categ in db.Categories select categ;
@@ -94,10 +136,13 @@ namespace ArticlesApp.Controllers
         // Adaugarea articolului in baza de date
 
         [HttpPost]
+        [Authorize(Roles = "Editor,Admin")]
         public IActionResult New(Article article)
         {
             article.Date = DateTime.Now;
-            article.Categ = GetAllCategories();
+            article.UserId = _userManager.GetUserId(User);
+
+            Console.WriteLine(article.UserId);
 
             if(ModelState.IsValid) 
             { 
@@ -108,6 +153,7 @@ namespace ArticlesApp.Controllers
             }
             else
             {
+                article.Categ = GetAllCategories();
                 return View(article);
             }
         }
@@ -116,6 +162,8 @@ namespace ArticlesApp.Controllers
         // Categoria se selecteaza dintr-un dropdown
         // HttpGet implicit
         // Se afiseaza formularul impreuna cu datele aferente articolului din baza de date
+
+        [Authorize(Roles = "Editor,Admin")]
         public IActionResult Edit(int id)
         {
             Article article = db.Articles.Include("Category")
@@ -124,27 +172,45 @@ namespace ArticlesApp.Controllers
 
             article.Categ = GetAllCategories();
 
-            return View(article);
+            if(article.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
+            {
+                return View(article);
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra unui articol care nu va \r\napartine";
+                return RedirectToAction("Index");
+            }
         }
 
         // Se adauga articolul modificat in baza de date
+
         [HttpPost]
+        [Authorize(Roles = "Editor,Admin")]
         public IActionResult Edit(int id, Article requestArticle)
         {   
             Article article = db.Articles.Find(id);
-            article.Categ = GetAllCategories();
 
             if(ModelState.IsValid)
             {
-                article.Title = requestArticle.Title;
-                article.Content = requestArticle.Content;
-                article.Date = requestArticle.Date;
-                article.CategoryId = requestArticle.CategoryId;
-                db.SaveChanges();
+                if(article.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
+                {
+                    article.Title = requestArticle.Title;
+                    article.Content = requestArticle.Content;
+                    article.Date = requestArticle.Date;
+                    article.CategoryId = requestArticle.CategoryId;
+                    db.SaveChanges();
 
-                TempData["message"] = "Articolul a fost modificat";
+                    TempData["message"] = "Articolul a fost modificat";
 
-                return RedirectToAction("Index");
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra unui articol care nu va \r\napartine";
+                    return RedirectToAction("Index");
+                }
+
             }
             else
             {
@@ -155,15 +221,28 @@ namespace ArticlesApp.Controllers
 
         // Se sterge un articol din baza de date 
         [HttpPost]
+        [Authorize(Roles = "Editor,Admin")]
         public ActionResult Delete(int id)
         {
-            Article article = db.Articles.Find(id);
-            db.Articles.Remove(article);
-            db.SaveChanges();
-            TempData["message"] = "Articolul a fost sters";
-            return RedirectToAction("Index");
+            Article article = db.Articles.Include("Comments")
+                                         .Where(art => art.Id == id)
+                                         .First();
+
+            if (article.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
+            {
+                db.Articles.Remove(article);
+                db.SaveChanges();
+                TempData["message"] = "Articolul a fost sters";
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa stergeti un articol care nu va apartine";
+                return RedirectToAction("Index");
+            }
         }
 
+        [NonAction]
         public IEnumerable<SelectListItem> GetAllCategories()
         {
             // generam o lista de tipul SelectListItem fara elemente
