+++ ArticlesApp/Controllers/CommentsController.cs
@@ -1,5 +1,7 @@
 ﻿using ArticlesApp.Data;
 using ArticlesApp.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 
 namespace ArticlesApp.Controllers
@@ -7,43 +9,81 @@ namespace ArticlesApp.Controllers
     public class CommentsController : Controller
     {
         private readonly ApplicationDbContext db;
-        public CommentsController(ApplicationDbContext context)
+
+        private readonly UserManager<ApplicationUser> _userManager;
+
+        private readonly RoleManager<IdentityRole> _roleManager;
+
+        public CommentsController(
+            ApplicationDbContext context,
+            UserManager<ApplicationUser> userManager,
+            RoleManager<IdentityRole> roleManager)
         {
             db = context;
+            _userManager = userManager;
+            _roleManager = roleManager;
         }
 
         // Stergerea unui comentariu asociat unui articol din baza de date
         [HttpPost]
+        [Authorize(Roles = "User,Editor,Admin")]
         public IActionResult Delete(int id)
         {
             Comment comm = db.Comments.Find(id);
-            db.Comments.Remove(comm);
-            db.SaveChanges();
-            return Redirect("/Articles/Show/" + comm.ArticleId);
+
+            if(comm.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
+            {
+                db.Comments.Remove(comm);
+                db.SaveChanges();
+                return Redirect("/Articles/Show/" + comm.ArticleId);
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa stergeti comentariul!";
+                return RedirectToAction("Index", "Articles");
+            }
         }
 
         // In acest moment vom implementa editarea intr-o pagina View separata
         // Se editeaza un comentariu existent
 
+        [Authorize(Roles = "User,Editor,Admin")]
         public IActionResult Edit(int id)
         {
             Comment comm = db.Comments.Find(id);
 
-            return View(comm);
+            if(comm.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
+            {
+                return View(comm);
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa editati comentariul!";
+                return RedirectToAction("Index", "Articles");
+            }
         }
 
-        [HttpPost]
+            [HttpPost]
+        [Authorize(Roles = "User,Editor,Admin")]
         public IActionResult Edit(int id, Comment requestComment)
         {
             Comment comm = db.Comments.Find(id);
             
             if(ModelState.IsValid)
             {
-                comm.Content = requestComment.Content;
+                if (comm.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
+                {
+                    comm.Content = requestComment.Content;
 
-                db.SaveChanges();
+                    db.SaveChanges();
 
-                return Redirect("/Articles/Show/" + comm.ArticleId);
+                    return Redirect("/Articles/Show/" + comm.ArticleId);
+                }
+                else
+                {
+                    TempData["message"] = "Nu aveti dreptul sa editati comentariul!";
+                    return RedirectToAction("Index", "Articles");
+                }
             }
             else
             {
